+++ internal/api/server.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	"allaboutapps.dev/aw/go-starter/internal/config"
+	"allaboutapps.dev/aw/go-starter/internal/i18n"
 	"allaboutapps.dev/aw/go-starter/internal/mailer"
 	"allaboutapps.dev/aw/go-starter/internal/mailer/transport"
 	"allaboutapps.dev/aw/go-starter/internal/push"
@@ -33,6 +34,7 @@ type Server struct {
 	Router *Router
 	Mailer *mailer.Mailer
 	Push   *push.Service
+	I18n   *i18n.Service
 }
 
 func NewServer(config config.Server) *Server {
@@ -43,6 +45,7 @@ func NewServer(config config.Server) *Server {
 		Router: nil,
 		Mailer: nil,
 		Push:   nil,
+		I18n:   nil,
 	}
 
 	return s
@@ -53,7 +56,8 @@ func (s *Server) Ready() bool {
 		s.Echo != nil &&
 		s.Router != nil &&
 		s.Mailer != nil &&
-		s.Push != nil
+		s.Push != nil &&
+		s.I18n != nil
 }
 
 func (s *Server) InitDB(ctx context.Context) error {
@@ -115,6 +119,18 @@ func (s *Server) InitPush() error {
 	if s.Push.GetProviderCount() < 1 {
 		log.Warn().Msg("No providers registered for push service")
 	}
+
+	return nil
+}
+
+func (s *Server) InitI18n() error {
+	i18nService, err := i18n.New(s.Config.I18n)
+
+	if err != nil {
+		return err
+	}
+
+	s.I18n = i18nService
 
 	return nil
 }